# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  errs.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.GetTeamBoards:
    description: Содержит данные о досках, принадлежащих команде
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.GetTeamResponse:
    description: Содержит данные о команде, включая пользователей и доски
    properties:
      boards:
        items:
          $ref: '#/definitions/models.GetTeamBoards'
        type: array
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      users:
        items:
          $ref: '#/definitions/models.GetTeamUsers'
        type: array
    type: object
  models.GetTeamUsers:
    description: Содержит данные о пользователях, принадлежащих команде
    properties:
      avatarUrl:
        type: string
      description:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления задачами и досками
  title: Avito FE Tech Internship 2025 Wave 2 API
  version: "1.0"
paths:
...
  /teams/{teamId}:
    get:
      consumes:
      - application/json
      description: Получает информацию о команде по ID, включая пользователей и доски
      parameters:
      - description: ID команды
        in: path
        name: teamId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetTeamResponse'
        "400":
          description: Некорректный teamID
          schema:
            $ref: '#/definitions/errs.ErrorResponse'
        "404":
          description: Команда не найдена
          schema:
            $ref: '#/definitions/errs.ErrorResponse'
      summary: Получить информацию о команде
      tags:
      - "Команды"
...
swagger: "2.0"